Add files via upload
Deployed a CNN for object recognition following a course from Eduonix on udemy. We used a All-CNN which was published by ICLR in 2015 on this particular link:
https://arxiv.org/pdf/1412.6806.pdf
This CNN obtained state-of-the-art performance at object recognition on the CIFAR-10 image dataset (2015). Also used a theao back-end instead of the tensorflow.
The dataset is obtained from https://www.cs.toronto.edu/~kriz/cifar.html or CIFAR-10 dataset, which has 60,000 (32,32,3) images in 10 classes. 50,000 are used for training and 10,000 for testing. As to save time, I have used the All CNN weights from https://github.com/PAN001/All-CNN/blob/master/all_cnn_weights_0.9088_0.4994.hdf5, which has pre-trained weights. 

The accuracy obtained by human on CIFAR-10 dataset was 94%, while this model tries to achieve 90%+ as well.